corrected the types in the meleeAttack function

+++ src/game/g_weapon.c
@@ -298,13 +298,13 @@ void meleeAttack( gentity_t *ent, float range, float width, float height,
 {
   trace_t   tr;
   gentity_t *traceEnt;
-  int n;
+  float n;
   float widthAdjusted, heightAdjusted;
 
   for( n=0; n <= 5; ++n )
   {
-    widthAdjusted = ( width * n ) / 5;
-    heightAdjusted = ( height * n ) / 5;
+    widthAdjusted = ( width * n ) / 5.00f;
+    heightAdjusted = ( height * n ) / 5.00f;
 
     G_WideTrace( &tr, ent, range, widthAdjusted, heightAdjusted, &traceEnt );
     if( traceEnt != NULL && traceEnt->takedamage )
